refactor(automation): address PR #1359 review — SRP, DRY, and cleanup fixes (#1289)
- Remove unused `import pytest` from test_post_merge_helpers.py (ruff F401)
- Extract `_pr_is_failing` + `FAILING_CHECK_CONCLUSIONS` to new leaf module
  `ci_predicates.py`; update ci_driver, post_merge_processor, pr_discovery to
  import from single source of truth (DRY)
- Remove redundant `self._pr_discovery._shared_pr_issues_setter = _spi_setter`
  re-assignment in CIDriver.__init__ (constructor already stored it)
- Move `_sweep_orphaned_arming_records` from CICheckInspector to
  PostMergeProcessor where it belongs by domain (SRP/POLA); wire `_gh_pr_state_fn`
  and `_state_dir` on PostMergeProcessor; update CIDriver delegation stub and
  all affected test patches
