+++ hephaestus/automation/ci_check_inspector.py
@@ -53,7 +53,7 @@ class CICheckInspector:
         self._status_tracker_update_slot = status_tracker_update_slot
 
         # Wired by CIDriver after construction — delegates to CIDriver methods
-        # that remain on the god class (arming store, post-merge /learn, etc.).
+        # that remain on the god class (arming store, per-issue arming check).
         self._load_arming_state_fn: Any = None
         self._clear_arming_state_fn: Any = None
         self._learn_record_terminal_fn: Any = None
@@ -258,83 +258,6 @@ class CICheckInspector:
             time.sleep(sleep_secs)
             elapsed += sleep_secs
             attempt += 1
-
-    # ------------------------------------------------------------------
-    # Arming-record sweep
-    # ------------------------------------------------------------------
-
-    def _sweep_orphaned_arming_records(self) -> None:
-        """Drop CLOSED records and capture missed ``/learn`` for MERGED orphans (#848).
-
-        The per-issue ``_check_arming_on_drive_start`` only fires when the
-        issue is in the current run's input list. If a PR was replaced via
-        the fresh-branch-reopen workflow (the replacement PR merged out of
-        band of the bot, the original was closed), the arming record points
-        at the closed PR and the issue itself is closed — the next run's
-        ``gh issue list --state open`` won't include it, so the record
-        leaks forever and ``/learn`` is silently lost.
-
-        Sweep every ``drive-green-armed-*.json`` at startup: drop records
-        whose PR is CLOSED-not-merged, fire ``/learn`` once for records
-        whose PR is MERGED (then mark explicit succeeded/failed learn status),
-        leave OPEN records alone for the normal per-issue path to handle.
-        """
-        state_dir = self._state_dir
-        if state_dir is None:
-            logger.info("Arming sweep skipped: state_dir not wired")
-            return
-        try:
-            records = sorted(state_dir.glob("drive-green-armed-*.json"))
-        except OSError as exc:
-            logger.info("Arming sweep skipped: state_dir scan failed (%s)", exc)
-            return
-        if not records:
-            return
-        logger.info("Sweeping %s arming record(s) for orphan resolution", len(records))
-        for path in records:
-            stem = path.stem  # drive-green-armed-<issue>
-            try:
-                issue_number = int(stem.rsplit("-", 1)[-1])
-            except ValueError:
-                logger.info("Arming sweep: ignoring malformed filename %s", path.name)
-                continue
-            record = self._load_arming_state_fn(issue_number)
-            if record is None:
-                continue
-            if self._learn_record_terminal_fn(record):
-                continue
-            pr_number = record.get("pr_number")
-            if not isinstance(pr_number, int):
-                logger.info(
-                    "Arming sweep: dropping record %s with non-integer pr_number",
-                    path.name,
-                )
-                self._clear_arming_state_fn(issue_number)
-                continue
-            gh_state = self._gh_pr_state(pr_number)
-            if gh_state is None:
-                continue
-            state = (gh_state.get("state") or "").upper()
-            if state == "MERGED":
-                logger.info(
-                    "Arming sweep: issue #%s / PR #%s MERGED; firing /learn",
-                    issue_number,
-                    pr_number,
-                )
-                learn_succeeded = self._run_drive_green_learnings_fn(issue_number, pr_number)
-                self._run_drive_green_compact_fn(issue_number, pr_number)
-                self._mark_drive_green_learn_result_fn(
-                    issue_number,
-                    record,
-                    succeeded=learn_succeeded,
-                )
-            elif state == "CLOSED":
-                logger.info(
-                    "Arming sweep: issue #%s / PR #%s CLOSED-not-merged; dropping record",
-                    issue_number,
-                    pr_number,
-                )
-                self._clear_arming_state_fn(issue_number)
 
     # ------------------------------------------------------------------
     # Per-issue arming check
