+++ tests/unit/automation/test_post_merge_helpers.py
@@ -5,8 +5,6 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Any
 
-import pytest
-
 from hephaestus.automation.post_merge_processor import PostMergeProcessor
